examples/email-service/requirements.txt
django==5.0.4
djangorestframework==3.15.1
requests==2.31.0
python-dotenv==1.0.1
